plugins/scope-moondream/pyproject.toml
[project]
name = "scope-moondream"
version = "0.1.0"
description = "Moondream vision language model plugin for Daydream Scope"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.51.1",
    "accelerate>=1.10.0",
    "Pillow>=11.0.0",
]

[project.entry-points."scope"]
scope_moondream = "scope_moondream"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scope_moondream"]
